backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
fastembed==0.6.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.2
groq==0.24.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
humanfriendly==10.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_pydantic==0.6
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-groq==0.3.2
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
loguru==0.7.3
marshmallow==3.26.1
mcp==1.8.0
mcp-use==1.2.10
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
onnxruntime==1.22.0
openai==1.78.0
orjson==3.10.18
packaging==24.2
pillow==11.2.1
propcache==0.3.1
protobuf==6.30.2
py_rust_stemmers==0.1.5
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.4
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
yarl==1.20.0
zstandard==0.23.0
